refacotr: 미사용 파일 삭제, resizeBlock 훅 리팩토링
